Refactor : Week4 부분 전체 리팩토링
